% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_generate.R
\name{build_doi_confirmation}
\alias{build_doi_confirmation}
\title{Determine what confirmation message (if any) should be shown before
creating a DataStore draft reference. Pure function - no side effects.}
\usage{
build_doi_confirmation(existing_doi_state)
}
\arguments{
\item{existing_doi_state}{NULL, or a list(reference_id = <7-digit int>,
doi = \if{html}{\out{<character>}}) from a prior successful creation this session}
}
\value{
list(needs_confirmation = TRUE, title = ..., message = ...)
}
\description{
Determine what confirmation message (if any) should be shown before
creating a DataStore draft reference. Pure function - no side effects.
}
