% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_generate.R
\name{build_generation_script}
\alias{build_generation_script}
\title{Assemble the full reproducible R script from every tab's state, in
skeleton.Rmd order. Pure function. Paths in the emitted script reference
the data_package_creation working directory, since that's where
EMLassemblyline's template functions and make_eml() actually operate.}
\usage{
build_generation_script(
  high_level_state,
  people_state,
  tables_state,
  fields_state,
  geo_state,
  taxonomy_state,
  working_folder_var = "working_folder"
)
}
\arguments{
\item{high_level_state}{the list returned by highLevelServer()'s reactive}

\item{people_state}{the list returned by peopleServer()'s reactive}

\item{tables_state}{the list returned by tableMetadataServer()'s reactive}

\item{fields_state}{the list returned by fieldsServer()'s reactive}

\item{geo_state}{the list returned by geographyServer()'s reactive}

\item{taxonomy_state}{the list returned by taxonomyServer()'s reactive}

\item{working_folder_var}{name of the R variable holding the working
folder path in the generated script (default "working_folder")}
}
\value{
character - the full generated R script as a single string
}
\description{
Assemble the full reproducible R script from every tab's state, in
skeleton.Rmd order. Pure function. Paths in the emitted script reference
the data_package_creation working directory, since that's where
EMLassemblyline's template functions and make_eml() actually operate.
}
