% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_generate.R
\name{derive_package_folder_name}
\alias{derive_package_folder_name}
\title{Derive the package folder name. Currently just the metadata_id - a
placeholder until enough package info exists to generate something more
descriptive (e.g. incorporating park unit + year).}
\usage{
derive_package_folder_name(high_level_state)
}
\arguments{
\item{high_level_state}{the list returned by highLevelServer()'s
reactive, evaluated (i.e. state <- high_level_reactive())}
}
\value{
character - the folder name to use for this package
}
\description{
Derive the package folder name. Currently just the metadata_id - a
placeholder until enough package info exists to generate something more
descriptive (e.g. incorporating park unit + year).
}
