% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_taxonomy.R
\name{emit_taxonomy_chunk}
\alias{emit_taxonomy_chunk}
\title{Emit the R code chunk for taxonomic coverage. Pure function.}
\usage{
emit_taxonomy_chunk(state, working_folder_var = "working_folder")
}
\arguments{
\item{state}{the list returned by taxonomyServer()'s reactive, evaluated
(i.e. state <- taxonomy_reactive())}

\item{working_folder_var}{name of the R variable holding the working
folder path in the generated script (default "working_folder")}
}
\value{
character - the R code chunk for taxonomic coverage templating,
or an explanatory comment if taxonomy coverage isn't enabled
}
\description{
Emit the R code chunk for taxonomic coverage. Pure function.
}
