% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fields.R
\name{validate_table_fields}
\alias{validate_table_fields}
\title{Validate one table's attributes + catvars state, per skeleton.Rmd/EML
schema requirements. Pure function - no side effects.}
\usage{
validate_table_fields(file_name, tbl_state)
}
\arguments{
\item{file_name}{the table's file name, used to prefix error messages
so the user knows which table/tab to go fix}

\item{tbl_state}{list(attributes = tibble, catvars = tibble) for one table}
}
\value{
character vector of error messages, empty if this table is valid
}
\description{
Validate one table's attributes + catvars state, per skeleton.Rmd/EML
schema requirements. Pure function - no side effects.
}
